requirements.txt
google-generativeai
pymupdf
scikit-learn
requests